requirements.txt
gdown>=5.2.0
matplotlib>=3.8.4
numpy>=1.24.3
opencv_python>=4.9.0.80
optuna>=3.5.0
pandas>=2.0.3
Pillow>=11.0.0
pytz>=2023.3.post1
requests>=2.32.3
scikit-learn>=1.3.0
torch>=2.2.1
torchinfo>=1.8.0
torchvision>=0.17.1
tqdm>=4.65.0
transformers>=4.40.0
